Core: Add experimental PTY support
Some processes change their behaviour when stdout/stderr are pipes or
flat out refuse to run when not connected to a terminal. This commit
adds basic support for using PTYs instead of pipes. It contains neither
correct handling if $TERM and terminfo nor correct chowning of the pipe
when dropping privileges.